% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal_Journal.R
\name{CreateNewEntry}
\alias{CreateNewEntry}
\title{Create dataframe holding test results}
\usage{
CreateNewEntry(script, times, warmup, hash)
}
\arguments{
\item{script}{Name of the script (with .R extension)}

\item{times}{Number of times to benchmark script (not including warmup)}

\item{warmup}{Number of times to run script first as a warmup}

\item{hash}{commit hash. Intended to be used only when using git hooks}
}
\value{
A dataframe with columns for the script name, the time taken to run
  the script, the package version (or commit hash), the version of R, the
  date of the test, and the name of the system
}
\description{
Runs a script multiple times and returns the benchmark results
  for the runs.
}
