% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal_Journal.R
\name{GetPkgName}
\alias{GetPkgName}
\title{Find name of current package}
\usage{
GetPkgName()
}
\value{
String. The name of the package being developed
}
\description{
Finds name of package being developed for by finding the current
    working directory and using the lowest level of the directory
}
