% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/User_Journal.R
\name{Init}
\alias{Init}
\title{Create folders in /man required for BenchJournal}
\usage{
Init()
}
\description{
Sets up a \code{BenchmarkJournal} subdirectory in the
    \code{/man} directory. Inside this subdirectory, a csv file is created to
    hold the benchmark results and two additional folders. One for holding
    scripts to be benchmarked, called \code{scripts}. and one to hold plots,
    helpfully called \code{plots}.The \code{BenchmarkJournal} subdirectory is
    also added to .Rbuildignore
}
