% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{JournalBoxPlot}
\alias{JournalBoxPlot}
\title{Create interactive boxplots of benchmark results from journal.}
\usage{
JournalBoxPlot(units = "s")
}
\arguments{
\item{units}{The unit of time to use for the plot. Either 's', 'ms', 'us'
(microseconds), or 'ns'.}
}
\description{
Uses \code{ggplot2} and \code{plotly} to create boxplots for
each script in \code{man/BenchJournal/scripts} by package version.
}
