% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/User_Journal.R
\name{NewEntry}
\alias{NewEntry}
\title{Create new entry of benchmark results in the journal.}
\usage{
NewEntry(times = 10)
}
\arguments{
\item{times}{The number of times to benchmark each script.}
}
\description{
Uses \code{microbenchmark} to benchmark the scripts in
man/BenchJournal/scripts and then adds the results to the journal file.
}
