% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/User_Journal.R
\name{NewJournal}
\alias{NewJournal}
\title{Create a new journal file}
\usage{
NewJournal(ask = TRUE)
}
\arguments{
\item{ask}{Logical. If TRUE then the user is asked to confirm overwriting
the journal csv file.}
}
\description{
Creates a new csv file to journal benchmark results, overwriting
the previous results. The user is asked to confirm overwriting an existing
file unless ask != TRUE
}
