% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal_Journal.R
\name{RemoveOutliers}
\alias{RemoveOutliers}
\title{Remove outliers from journal}
\usage{
RemoveOutliers(df)
}
\arguments{
\item{df}{dataframe in the format of journal}
}
\value{
A dataframe where observations with outlier run times have been
  removed
}
\description{
Remove observations from a journal with outlier run times.
}
